# openapi.yaml
openapi: 3.0.3
info:
  title: TaskMaster API
  description: |
    **TaskMaster Enhanced REST API v2 with Real-time Features**
    
    TaskMaster is a comprehensive task and schedule management platform featuring:
    - Enhanced REST API v2 with pagination, filtering, and sorting
    - Real-time WebSocket integration for live updates
    - Advanced search with full-text search and faceting
    - Data export in multiple formats
    - Webhook notifications for external integrations
    - AI-powered task analysis and recommendations
    - Performance monitoring and analytics
    
    ## Key Features
    
    ### Enhanced REST API v2
    - **Pagination**: Consistent pagination across all list endpoints
    - **Filtering**: Advanced filtering with date ranges, status, priority
    - **Sorting**: Multi-field sorting with direction control
    - **Field Selection**: Reduce payload size by selecting specific fields
    - **Batch Operations**: Create/update multiple resources efficiently
    - **Conditional Requests**: ETag and Last-Modified support
    
    ### Real-time Features
    - **WebSocket Integration**: Live updates for tasks, events, and schedules
    - **Event Broadcasting**: Real-time notifications across connected clients
    - **AI Insights**: Instant delivery of analysis results
    - **Collaborative Features**: Multi-user synchronization
    
    ### Advanced Search
    - **Full-Text Search**: Search across all entity types
...
    
    ### Data Management
...
    
...
    
...
    
...
servers:
...
paths:
...
  /v2/tasks:
    get:
      tags: ["Tasks v2"]
      summary: List Tasks with Advanced Filtering
      description: |
        Get tasks with comprehensive filtering, sorting, and pagination.
        Supports full-text search across title and description fields.
      operationId: listTasksV2
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Fields'
        - name: status
          in: query
          description: Filter by task status
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TaskStatus'
          style: form
          explode: false
        - name: priority
          in: query
          description: Filter by priority levels
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Priority'
          style: form
          explode: false
        - name: is_completed
          in: query
          description: Filter by completion status
          schema:
            type: boolean
        - name: is_snoozed
          in: query
          description: Filter by snooze status
          schema:
            type: boolean
        - name: due_date_from
          in: query
          description: Filter tasks due after this date
          schema:
            type: string
            format: date
        - name: due_date_to
          in: query
          description: Filter tasks due before this date
          schema:
            type: string
            format: date
        - name: created_from
          in: query
          description: Filter tasks created after this datetime
          schema:
            type: string
            format: date-time
        - name: created_to
          in: query
          description: Filter tasks created before this datetime
          schema:
            type: string
            format: date-time
        - name: duration_min
          in: query
          description: Minimum task duration in minutes
          schema:
            type: integer
            minimum: 1
        - name: duration_max
          in: query
          description: Maximum task duration in minutes
          schema:
            type: integer
            minimum: 1
        - name: urgency_min
          in: query
          description: Minimum urgency level
          schema:
            type: integer
            minimum: 1
            maximum: 10
        - name: urgency_max
          in: query
          description: Maximum urgency level
          schema:
            type: integer
            minimum: 1
            maximum: 10
        - name: initiative_id
          in: query
          description: Filter by initiative ID
          schema:
            type: string
            format: uuid
        - name: project_id
          in: query
          description: Filter by project ID
          schema:
            type: string
            format: uuid
        - name: parent_task_id
          in: query
          description: Filter by parent task ID
          schema:
            type: string
            format: uuid
        - name: root_only
          in: query
          description: Only show tasks with no incomplete dependencies
          schema:
            type: boolean
      responses:
        '200':
          description: List of tasks retrieved successfully
          headers:
            X-Total-Count:
              description: Total number of tasks matching criteria
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: ["Tasks v2"]
      summary: Create New Task
      description: Create a new task with validation and automatic field population
      operationId: createTaskV2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          headers:
            Location:
              description: URL of the created task
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication for programmatic access.
        Required for analytics, export, and webhook endpoints.
        
        API keys can be scoped with specific permissions:
        - `tasks.read`, `tasks.write` - Task operations
        - `events.read`, `events.write` - Event operations
        - `search.read` - Search operations
        - `export.create` - Data export
        - `webhooks.read`, `webhooks.create` - Webhook management
        - `analytics.read` - Usage analytics
    
    SessionAuth:
      type: http
      scheme: bearer
      description: |
        Session-based authentication for web applications.
        Obtained through login endpoints.
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication for mobile and SPA applications.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 20
      example: 50
    Offset:
      name: offset
      in: query
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 100
    Sort:
      name: sort
      in: query
      description: Sort order in format 'field:direction,field2:direction'
      schema:
        type: string
        pattern: '^(\w+:(asc|desc))(,\w+:(asc|desc))*$'
      example: "priority:desc,created_at:asc"
    SearchQuery:
      name: q
      in: query
      description: Full-text search query
      schema:
        type: string
        maxLength: 500
      example: "project management urgent"
    Fields:
      name: fields
      in: query
      description: Comma-separated list of fields to include in response
      schema:
        type: string
      example: "id,title,status,priority"
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          readOnly: true
        title:
          type: string
          description: Task title
          maxLength: 200
          example: "Complete project documentation"
        description:
          type: string
          description: Detailed task description
          example: "Write comprehensive API documentation including examples"
        duration:
          type: integer
          description: Estimated duration in minutes
          minimum: 1
          maximum: 4320
          example: 120
        urgency:
          type: integer
          description: Urgency level (1-10, where 10 is most urgent)
          minimum: 1
          maximum: 10
          example: 7
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        due_date:
          type: string
          format: date
          description: Task due date
          example: "2025-09-25"
        due_time:
          type: string
          format: time
          description: Specific due time
          example: "14:30:00"
        completed:
          type: boolean
          description: Task completion status
          readOnly: true
        is_divisible:
          type: boolean
          description: Whether task can be split into smaller chunks
          default: false
        min_chunk_size:
          type: integer
          description: Minimum chunk size if divisible (minutes)
          minimum: 5
        required_weather:
          type: array
          items:
            type: string
            enum: ["sunny", "cloudy", "rainy", "snowy", "indoor"]
          description: Weather conditions required for task
        required_context:
          type: array
          items:
            type: string
          description: Context tags required for task execution
          example: ["office", "computer", "internet"]
        equipment_needed:
          type: array
          items:
            type: string
          description: Equipment required for task
          example: ["laptop", "external monitor"]
        depends_on_task_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Task IDs this task depends on
        blocks_task_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Task IDs this task blocks
        initiative_id:
          type: string
          format: uuid
          description: Associated initiative ID
        project_id:
          type: string
          format: uuid
          description: Associated project ID
        parent_task_id:
          type: string
          format: uuid
          description: Parent task ID for subtasks
        is_recurring:
          type: boolean
          description: Whether task is recurring
          readOnly: true
        recurrence_days:
          type: integer
          description: Days between recurring instances
          minimum: 1
        is_snoozed:
          type: boolean
          description: Whether task is currently snoozed
          readOnly: true
        snoozed_until:
          type: string
          format: date-time
          description: When snooze expires
          readOnly: true
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          readOnly: true
      required:
        - title
    TaskCreateRequest:
      type: object
      properties:
        title:
          type: string
          description: Task title
          maxLength: 200
          minLength: 1
        description:
          type: string
          description: Task description
          maxLength: 5000
        duration:
          type: integer
          description: Estimated duration in minutes
          minimum: 1
          maximum: 4320
          default: 30
        urgency:
          type: integer
          description: Urgency level (1-10)
          minimum: 1
          maximum: 10
          default: 5
        priority:
          $ref: '#/components/schemas/Priority'
        due_date:
          type: string
          format: date
          description: Task due date
        start_date:
          type: string
          format: date-time
          description: Earliest start time
        required_weather:
          type: array
          items:
            type: string
            enum: ["sunny", "cloudy", "rainy", "snowy", "indoor"]
        initiative_id:
          type: string
          format: uuid
          description: Associated initiative
        project_id:
          type: string
          format: uuid
          description: Associated project
        parent_task_id:
          type: string
          format: uuid
          description: Parent task for subtasks
        recurrence_days:
          type: integer
          description: Days between recurring instances
          minimum: 1
      required:
        - title
      example:
        title: "Review and approve pull request"
        description: "Review code changes for new feature implementation"
        duration: 45
        urgency: 6
        priority: "high"
        due_date: "2025-09-22"
        project_id: "123e4567-e89b-12d3-a456-426614174000"
...
    TaskStatus:
      type: string
      enum: ["active", "blocked", "completed", "snoozed"]
      description: |
        Task status:
        - `active`: Task is ready to be worked on
        - `blocked`: Task is waiting for dependencies
        - `completed`: Task has been finished
        - `snoozed`: Task is temporarily hidden
      example: "active"
    Priority:
      type: string
      enum: ["low", "medium", "high", "urgent"]
      description: |
        Priority levels:
        - `low`: Nice to have, no urgency
        - `medium`: Normal priority work
        - `high`: Important, should be done soon
        - `urgent`: Critical, needs immediate attention
      example: "high"

    # Event Schemas
...
    TaskResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["success"]
        data:
          $ref: '#/components/schemas/Task'
        message:
          type: string
    TaskListResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["success"]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        has_more:
          type: boolean
          description: Whether more items are available
        page:
          type: integer
          description: Current page number (1-indexed)
        total_pages:
          type: integer
          description: Total number of pages

    # Export Schemas
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["error"]
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        errors:
          type: object
          additionalProperties:
            type: string
          description: Field-specific validation errors
        details:
          type: string
          description: Additional error details
      required:
        - status
        - message
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Invalid request parameters"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Authentication required"
            code: "UNAUTHORIZED"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Validation failed"
            code: "VALIDATION_ERROR"
            errors:
              title: "Title is required"
              duration: "Duration must be positive"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Internal server error"
            code: "INTERNAL_ERROR"

# Global security requirement (can be overridden per operation)
security:
  - SessionAuth: []
  - ApiKeyAuth: []

# Tags for grouping operations
...